requirements.txt
beautifulsoup4==4.14.3
gitdb==4.0.12
GitPython==3.1.45
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
rich==14.2.0
semantic-version==2.10.0
smmap==5.0.2
soupsieve==2.8
types-beautifulsoup4==4.12.0.20250516
types-html5lib==1.1.11.20251117
types-webencodings==0.5.0.20251108
typing_extensions==4.15.0
